ls-0006 Move screen output function to a new file
